Automata

To represent the different control behaviors that can result from deadline misses, we make use of finite-state transducer automata. These take as input a sequence of small integers representing deadline hits and misses, or potentially other scheduler actions.

ControlTimingSafety.Automaton — Type

Automaton(Φ, T, μ, l_int)

A transducer automaton.

See also the constructors hold_kill, hold_skip_next, zero_kill, and zero_skip_next, which create Automaton objects for common deadline miss handling strategies. For use in notebooks, etc., see strat_map and strat_names.

ControlTimingSafety.Automaton — Method

Automaton(a, l_int)

Construct a copy of the Automaton a with a new l_int.

Constructors for Miss Handling Strategies

As a user of ControlTimingSafety, you will likely not want to construct an Automaton directly. The package provides external constructors that create an Automaton from a StateSpace model from ControlSystems.jl, following different strategies of how to handle deadline misses. Constraints may be specified using WeaklyHardConstraint objects from RealTimeScheduling.jl.

Note

Currently, only MissRow constraints are supported, but more constraints will be added in the future!

ControlTimingSafety.hold_kill — Function

hold_kill(sysd, K, [c])

Construct an Automaton for the given discrete-time state space model sysd with feedback gain matrix K, following the Hold&Kill strategy. If c is specified, the resulting Automaton will permit only that weakly hard constraint.

ControlTimingSafety.hold_skip_next — Function

hold_skip_next(sysd, K, [c])

Construct an Automaton for the given discrete-time state space model sysd with feedback gain matrix K, following the Hold&Skip-Next strategy. If c is specified, the resulting Automaton will permit only that weakly hard constraint.

ControlTimingSafety.zero_kill — Function

zero_kill(sysd, K, [c])

Construct an Automaton for the given discrete-time state space model sysd with feedback gain matrix K, following the Zero&Kill strategy. If c is specified, the resulting Automaton will permit only that weakly hard constraint.

ControlTimingSafety.zero_skip_next — Function

zero_skip_next(sysd, K, [c])

Construct an Automaton for the given discrete-time state space model sysd with feedback gain matrix K, following the Zero&Skip-Next strategy. If c is specified, the resulting Automaton will permit only that weakly hard constraint.

For convenience when building interfaces, e.g. notebooks, we also export a sorted list of human-readable names for these constructors, and a map from these names to the functions themselves.

ControlTimingSafety.strat_names — Constant

A stable, sorted list of the human-readable names in strat_map.

ControlTimingSafety.strat_map — Constant

A mapping from human-readable names to Automaton constructor functions.

See also strat_names.

Computing Evolution

Given an initial state z0 and a sequence of scheduler actions seq, the evolution of an Automaton can be computed. Since the augmented state space requires extra state variables outside the plant's state, the state z0 can be computed from a plant state x0 using the augment function.

ControlTimingSafety.augment — Function

augment(a::Automaton, x)

Pad the state vector (or matrix of column vectors) x to as many dimensions as in the augmented state space of a.

Once the initial augmented state is determined, we can then compute its evolution.

ControlTimingSafety.evol — Function

evol(a, z_0, input)

Simulate a run of Automaton a from the initial state z_0, using the sequence of scheduler events input.

Returns z, a matrix of states over time.

See also evol_final, which additionally returns the final location in the automaton.

ControlTimingSafety.evol_final — Function

evol_final(a, z_0, input)

Same as evol, but returns (z, l), where z is a matrix of states over time, and l is the final location in the automaton.

Missing docstring.

Missing docstring for evol!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for evol_final!. Check Documenter's build log for details.