Probable Safety

In the paper "Statistical Hypothesis Testing of Controller Implementations Under Timing Uncertainties," we developed a method for estimating the maximum deviation of a control system that may experience deadline misses. This estimate is based on Jeffreys' Bayes factor hypothesis testing, and is probabilistically guaranteed. The method is implemented in the estimate_deviation function.

ControlTimingSafety.estimate_deviation — Function

estimate_deviation(a, sampler, z_0, c, B; nominal, nominal_trajectory)

Estimate the deviation possible in the Automaton a, using sampler to generate random behaviors. The initial state is given as z_0. c specifies the confidence with which the returned value is asserted to be an upper bound, and B is the Bayes factor used in hypothesis testing. The parameters metric, nominal, and nominal_trajectory are as in deviation.

For more information, see Bineet Ghosh et al., "Statistical Hypothesis Testing of Controller Implementations Under Timing Uncertainties," RTCSA 2022. DOI: 10.1109/RTCSA55878.2022.00008

ControlTimingSafety.maximum_deviation_random — Function

maximum_deviation_random(a, sampler, z_0, samples; estimate, metric, nominal,
                         nominal_trajectory)

Calculate samples random behaviors using sampler, and the corresponding trajectories of the Automaton a from the initial state z_0. Return the maximum deviation from a nominal trajectory. The parameters metric, nominal, and nominal_trajectory are as in deviation. If estimate is specified, stop early if this deviation estimate is exceeded, returning the exceeding deviation.