Checking Safety

The main safety-checking function of ControlTimingSafety is bounded_runs_iter, which implements the Bounded Runs Iteration algorithm from "Safety Analysis of Embedded Controllers under Implementation Platform Timing Uncertainties." This implementation permits an initial point, or an axis-aligned box initial set. The helper function bounded_runs, which computes a single iteration, is also exported.

ControlTimingSafety.bounded_runs_iter — Function

bounded_runs_iter(a, z_0, n, t)

Iterate bounded_runs(a, z_0, n) for t iterations, returning the reachable set at each of the n×t+1 time steps.

See also deviation, which can be called with the result of this function to find the deviation from a nominal trajectory.

ControlTimingSafety.bounded_runs — Function

bounded_runs(a::Automaton, z_0, n)

Compute reachable sets for n time steps for the given Automaton a, starting from the initial set given by z_0.

z_0 must be an a.nz-element vector, or an a.nz×2 matrix whose first and second columns are the minimum and maximum along each dimension, respectively.

Returns (bounds, locs), where bounds is an nactions(a)^n×n+1×a.nz×2 Array{Float64} giving the bounding box for each (run, time step), and locs is an nactions(a)^n Array{Int64} of final locations for each run (or zero if there is no such run).

See also bounded_runs_iter, which calls this function iteratively to efficiently compute reachable sets for longer time horizons. Typically one will call deviation on the results of this function to determine deviation from a nominal trajectory.

After computing a list of reachable sets, safety can be determined by computing its distance from a nominal run. This is done using the deviation function.

ControlTimingSafety.deviation — Function

deviation(a, z_0, reachable; metric=Euclidean(), nominal=[1,1,...])

Compute the deviation from the nominal behavior (default: all 1) that is possible for the given Automaton a, starting from the set of initial states z_0, within the reachable sets. The deviation is computed using the specified metric from the Distances.jl package, on the output of the system (i.e. after multiplying by a.C). reachable may be the three-dimensional output of e.g. bounded_runs_iter, or a matrix with dimensions (state, time), e.g. from evol.

If nominal_trajectory is given, this trajectory is used instead of computing the trajectory from the nominal behavior. This can improve efficiency when calling deviation iteratively.

See also bounded_runs and bounded_runs_iter, which can be used to compute reachable.

Utility Functions

A few utility functions are exported as well. These support the implementation of the above functions, and are potentially useful to their callers as well.

ControlTimingSafety.corners_from_bounds — Function

corners_from_bounds(bounds::AbstractMatrix; cycle=false, dims=axes(bounds, 1))

Returns the corners of the n-dimensional interval represented by bounds. If cycle is true, the first corner is repeated at the end, and the corners are given in Gray code order. Only the dimensions from dims are considered.

corners_from_bounds(bounds::AbstractVector; cycle=nothing, dims=nothing)

When applied to a vector, cast it to a one-column matrix. cycle and dims are ignored.

ControlTimingSafety.merge_bounds — Function

merge_bounds(b)

Merges an array of bounding boxes b into one.